Complete operations for weibo and tieba

+++ src/main/java/com/woodpecker/domain/TableCollection.java
@@ -1,10 +1,10 @@
 package com.woodpecker.domain;
 
 public class TableCollection extends UserCollection {
-    private Integer tableid;
-    public Integer getTableid(){return tableid;}
-    public void setTableid(Integer tableid){this.tableid=tableid;}
-    public TableCollection(Integer dataid, String data, Integer tableid){
+    private Long tableid;
+    public Long getTableid(){return tableid;}
+    public void setTableid(Long tableid){this.tableid=tableid;}
+    public TableCollection(Integer dataid, String data, Long tableid){
         this.setDataid(dataid);
         this.setData(data);
         this.tableid=tableid;
